+++ crates/workspace/src/panels/messaging_relays.rs
@@ -150,7 +150,7 @@ impl MessagingRelayPanel {
         let value = self.input.read(cx).value().to_string();
 
         if !value.starts_with("ws") {
-            self.set_error("Relay URl is invalid", window, cx);
+            self.set_error("The relay URL is invalid.", window, cx);
             return;
         }
 
@@ -163,7 +163,7 @@ impl MessagingRelayPanel {
                 cx.notify();
             }
         } else {
-            self.set_error("Relay URl is invalid", window, cx);
+            self.set_error("The relay URL is invalid.", window, cx);
         }
     }
 
@@ -200,7 +200,7 @@ impl MessagingRelayPanel {
 
     pub fn set_relays(&mut self, window: &mut Window, cx: &mut Context<Self>) {
         if self.relays.is_empty() {
-            self.set_error("You need to add at least 1 relay", window, cx);
+            self.set_error("Add at least one relay.", window, cx);
             return;
         };